created new component file with functionality added for vote count
