--- src/components/VoteCount.jsx
@@ -1,3 +1,0 @@
-const VoteCount = () => {};
-
-export default VoteCount;
